+++ Plugins/TableProPluginKit/PluginRoutineInfo.swift
@@ -4,7 +4,7 @@ public struct PluginRoutineInfo: Codable, Sendable {
     public let name: String
     public let type: String
 
-    public init(name: String, type: String = "PROCEDURE") {
+    public init(name: String, type: String) {
         self.name = name
         self.type = type
     }