+++ TablePro/Models/Schema/RoutineInfo.swift
@@ -1,11 +1,11 @@
 import Foundation
 
 struct RoutineInfo: Identifiable, Hashable {
-    var id: String { "\(name)_\(type.rawValue)" }
+    var id: String { name }
     let name: String
     let type: RoutineType
 
-    enum RoutineType: String, Sendable {
+    enum RoutineType: String, Codable, Sendable {
         case procedure = "PROCEDURE"
         case function = "FUNCTION"
     }